requirements.txt
pyobjc-core
pyobjc
pyautogui
pillow
pyqt5
msgpack-python
redis

brew install redis